fix the bloody time

+++ twoOpt.h
@@ -6,5 +6,5 @@ using namespace std;
 using namespace std;
 vector<int> opt2(Cities &cities, vector<int> tour);
 vector<int> swap(vector<int> tour, int i, int k );
-vector<int> twoOpt(Cities &cities, float time);
+vector<int> twoOpt(Cities &cities, double time);
 #endif
